+++ src/ignition_cells.hpp
@@ -3,6 +3,6 @@
 #include <string>
 #include <vector>
 
-typedef std::vector<std::pair<unsigned int, unsigned int>> IgnitionCells;
+typedef std::vector<std::pair<size_t, size_t>> IgnitionCells;
 
 IgnitionCells read_ignition_cells(std::string filename);
