+++ src/spread_functions.cpp
@@ -10,16 +10,16 @@
 #include "fires.hpp"
 #include "landscape.hpp"
 
-float spread_probability(
-    const Cell& burning, const Cell& neighbour, SimulationParams params, float angle,
-    float distance, float elevation_mean, float elevation_sd, float upper_limit = 1.0
+double spread_probability(
+    const Cell& burning, const Cell& neighbour, SimulationParams params, double angle,
+    double distance, double elevation_mean, double elevation_sd, double upper_limit = 1.0
 ) {
 
-  float slope_term = sin(atan((neighbour.elevation - burning.elevation) / distance));
-  float wind_term = cos(angle - burning.wind_direction);
-  float elev_term = (neighbour.elevation - elevation_mean) / elevation_sd;
+  double slope_term = sin(atan((neighbour.elevation - burning.elevation) / distance));
+  double wind_term = cos(angle - burning.wind_direction);
+  double elev_term = (neighbour.elevation - elevation_mean) / elevation_sd;
 
-  float linpred = params.independent_pred;
+  double linpred = params.independent_pred;
 
   if (neighbour.vegetation_type == SUBALPINE) {
     linpred += params.subalpine_pred;
@@ -35,53 +35,53 @@ float spread_probability(
   linpred += wind_term * params.wind_pred + elev_term * params.elevation_pred +
              slope_term * params.slope_pred;
 
-             float prob = upper_limit / (1 + exp(-linpred));
+             double prob = upper_limit / (1 + exp(-linpred));
 
   return prob;
 }
 
 Fire simulate_fire(
-    const Landscape& landscape, const std::vector<std::pair<unsigned int, unsigned int>>& ignition_cells,
-    SimulationParams params, float distance, float elevation_mean, float elevation_sd,
-    float upper_limit = 1.0
+    const Landscape& landscape, const std::vector<std::pair<size_t, size_t>>& ignition_cells,
+    SimulationParams params, double distance, double elevation_mean, double elevation_sd,
+    double upper_limit = 1.0
 ) {
 
-  unsigned int n_row = landscape.height;
-  unsigned int n_col = landscape.width;
+  size_t n_row = landscape.height;
+  size_t n_col = landscape.width;
 
-  std::vector<std::pair<unsigned int, unsigned int>> burned_ids;
+  std::vector<std::pair<size_t, size_t>> burned_ids;
 
-  unsigned int start = 0;
-  unsigned int end = ignition_cells.size();
+  size_t start = 0;
+  size_t end = ignition_cells.size();
 
-  for (unsigned int i = 0; i < end; i++) {
+  for (size_t i = 0; i < end; i++) {
     burned_ids.push_back(ignition_cells[i]);
   }
 
-  std::vector<unsigned int> burned_ids_steps;
+  std::vector<size_t> burned_ids_steps;
   burned_ids_steps.push_back(end);
 
-  unsigned int burning_size = end + 1;
+  size_t burning_size = end + 1;
 
   Matrix<bool> burned_bin = Matrix<bool>(n_col, n_row);
 
-  for (unsigned int i = 0; i < end; i++) {
-    unsigned int cell_0 = ignition_cells[i].first;
-    unsigned int cell_1 = ignition_cells[i].second;
+  for (size_t i = 0; i < end; i++) {
+    size_t cell_0 = ignition_cells[i].first;
+    size_t cell_1 = ignition_cells[i].second;
     burned_bin[{ cell_0, cell_1 }] = 1;
   }
 
   double t = omp_get_wtime();
   while (burning_size > 0) {
-    unsigned int end_forward = end;
+    size_t end_forward = end;
 
     // Loop over burning cells in the cycle
 
     // b is going to keep the position in burned_ids that have to be evaluated
     // in this burn cycle
-    for (unsigned int b = start; b < end; b++) {
-      unsigned int burning_cell_0 = burned_ids[b].first;
-      unsigned int burning_cell_1 = burned_ids[b].second;
+    for (size_t b = start; b < end; b++) {
+      size_t burning_cell_0 = burned_ids[b].first;
+      size_t burning_cell_1 = burned_ids[b].second;
 
       const Cell& burning_cell = landscape[{ burning_cell_0, burning_cell_1 }];
 
@@ -90,7 +90,7 @@ Fire simulate_fire(
 
       int neighbors_coords[2][8];
 
-      for (unsigned int i = 0; i < 8; i++) {
+      for (size_t i = 0; i < 8; i++) {
         neighbors_coords[0][i] = int(burning_cell_0) + moves[i][0];
         neighbors_coords[1][i] = int(burning_cell_1) + moves[i][1];
       }
@@ -98,7 +98,7 @@ Fire simulate_fire(
 
       // Loop over neighbors_coords of the focal burning cell
 
-      for (unsigned int n = 0; n < 8; n++) {
+      for (size_t n = 0; n < 8; n++) {
 
         int neighbour_cell_0 = neighbors_coords[0][n];
         int neighbour_cell_1 = neighbors_coords[1][n];
@@ -119,17 +119,17 @@ Fire simulate_fire(
         if (!burnable_cell)
           continue;
 
-        constexpr float angles[8] = { M_PI * 3 / 4, M_PI, M_PI * 5 / 4, M_PI / 2, M_PI * 3 / 2,
+        constexpr double angles[8] = { M_PI * 3 / 4, M_PI, M_PI * 5 / 4, M_PI / 2, M_PI * 3 / 2,
                                        M_PI / 4,     0,    M_PI * 7 / 4 };
 
         // simulate fire
-        float prob = spread_probability(
+        double prob = spread_probability(
             burning_cell, neighbour_cell, params, angles[n], distance, elevation_mean,
             elevation_sd, upper_limit
         );
 
         // Burn with probability prob (Bernoulli)
-        bool burn = (float)rand() / (float)RAND_MAX < prob;
+        bool burn = (double)rand() / (double)RAND_MAX < prob;
 
         if (burn == 0)
           continue;
